[backups] Implement a job strategy controller
## What this PR does

A new `Job.strategy.backups.cozystack.io` resource is implemented to
represent a backup strategy that uses a one-shot job to create and
upload a backup. An accompanying controller is also created to respond
to BackupJobs that are created with a reference to this strategy. This
implementation can be used as a reference by third-party app developers
to implement their own strategies and controllers.

### Release note

```release-note
[backups] Implement a job backup strategy and controller
```

Signed-off-by: Timofei Larkin <lllamnyp@gmail.com>