[backups] Fix malformed glob and split in template (#1708)
## What this PR does

A malformed glob ("*" instead of "*.yaml") captured the .gitattributes
file and sent it to the templating engine. Using `split` instead of
`splitList` on a string returned a map instead of a list, so templating
broke on `mustLast`. This patch corrects the errors.

### Release note

```release-note
[backups] Fix template-breaking errors in the backup-controller Helm
chart.
```

(cherry picked from commit 4f5ae287f59ec7195a9404a72eddba5004065043)
Signed-off-by: Timofei Larkin <lllamnyp@gmail.com>